+++ examples/custom_xmlrpc_client/xmlrpc_locustfile.py
@@ -7,7 +7,7 @@ from locust import User, task, between
 class XmlRpcClient(ServerProxy):
     """
     Simple, sample XML RPC client implementation that wraps xmlrpclib.ServerProxy and
-    fires locust events on request_success and request_failure, so that all requests
+    fires locust events on request, so that all requests
     gets tracked in locust's statistics.
     """
 
@@ -18,21 +18,22 @@ class XmlRpcClient(ServerProxy):
 
         def wrapper(*args, **kwargs):
             start_time = time.time()
-            exc = None
+            request_meta = {
+                "request_type": "xmlrpc",
+                "name": name,
+                "response_time": 0,
+                "response_length": 0,
+                "context": {},
+                "exception": None,
+            }
+
             try:
                 result = func(*args, **kwargs)
             except Fault as e:
-                exc = e
+                request_meta["exception"] = e
 
-            total_time = int((time.time() - start_time) * 1000)
-            self._locust_environment.events.request.fire(
-                request_type="xmlrpc",
-                name=name,
-                response_time=total_time,
-                response_length=0,
-                context={},
-                exception=exc,
-            )
+            request_meta["response_time"] = int((time.time() - start_time) * 1000)
+            self._locust_environment.events.request.fire(**request_meta)
             # In this example, I've hardcoded response_length=0. If we would want the response length to be
             # reported correctly in the statistics, we would probably need to hook in at a lower level