+++ docs/changelog.rst
@@ -8,7 +8,7 @@ For full details of the Locust changelog, please see https://github.com/locustio
 =====
 
 * Ensure runner.quit finishes even when users are broken https://github.com/locustio/locust/pull/1728
-* Make user count available to LoadTestShape https://github.com/locustio/locust/pull/1719
+* Make runner / user count available to LoadTestShape https://github.com/locustio/locust/pull/1719
 * Other small fixes
 
 1.4.3