+++ docs/changelog.rst
@@ -7,8 +7,8 @@ For full details of the Locust changelog, please see https://github.com/locustio
 1.5.0
 =====
 
-* Add new event called request. Is called on every request successful or not. request_success and request_failure are still available but can be considered deprecated
-* Add parameter to the request event. Can be used to forward information when calling a request, things like user information, tags etc
+* Add new event called request. Is called on every request successful or not. request_success and request_failure are still available but are deprecated
+* Add parameter context to the request event. Can be used to forward information when calling a request, things like user information, tags etc
 
 1.4.4
 =====